chore: updates jscong moduleResolution
